分出單筆軌跡
