+++ src/main/java/com/example/pentaho/service/SingleQueryService.java
@@ -198,11 +198,6 @@ public class SingleQueryService {
         return originalString;
     }
 
-    public List<IbdTbIhChangeDoorplateHis> singleQueryTrack(String addressId) {
-        log.info("addressId:{}",addressId);
-        return ibdTbIhChangeDoorplateHisRepository.findByAddressId(addressId);
-    }
-
 
     public String findNeighborCd(String rawNeighbor) {
         Pattern pattern = Pattern.compile("\\d+"); //指提取數字
@@ -217,219 +212,5 @@ public class SingleQueryService {
             log.info("沒有數字部分");
             return "";
         }
-    }
-
-
-    @Async
-    public void queryBatchTrack(MultipartFile file, SingleBatchQueryParams singleBatchQueryParams) throws IOException {
-        log.info("singleBatchQueryParams:{}",singleBatchQueryParams.toString());
-        try {
-            String filePath="";
-            String fileContent = new String(file.getBytes(), StandardCharsets.UTF_8);
-            int lastIndexOf = fileContent.lastIndexOf("\n");
-            fileContent = fileContent.substring(0,lastIndexOf-1);
-            log.info("fileContent:{}",fileContent);
-            String[] lines = fileContent.split("\n");
-            if(lines.length <= 0){
-                /**檔案沒有內容*/
-                singleBatchQueryParams.setStatus("REJECT");
-                postSingleBatchQueryRequest("/batchForm/systemUpdate",singleBatchQueryParams,filePath);
-                return;
-            }
-            String[] newLines = Arrays.copyOfRange(lines, 1, lines.length);
-            log.info("newLines:{}",newLines.length);
-            if(newLines.length <= 0){
-                singleBatchQueryParams.setStatus("REJECT");
-                postSingleBatchQueryRequest("/batchForm/systemUpdate",singleBatchQueryParams,filePath);
-                return;
-            }
-
-            /**表頭除外之筆數，空白行也算**/
-            singleBatchQueryParams.setProcessedCounts(String.valueOf(newLines.length));
-
-            /**取得地址編碼，記得略過表頭**/
-            List<String> addressIdList = CSVReader(file);
-            if(addressIdList == null){
-                singleBatchQueryParams.setStatus("SYS_FAILED");
-                postSingleBatchQueryRequest("/batchForm/systemUpdate",singleBatchQueryParams,filePath);
-                return;
-            }
-
-            List<IbdTbIhChangeDoorplateHis> IbdTbIhChangeDoorplateHisList = ibdTbIhChangeDoorplateHisRepository.findByAddressIdList(addressIdList);
-            boolean geneFile = getCSVFile(IbdTbIhChangeDoorplateHisList, singleBatchQueryParams.getFile());
-            if(geneFile){
-                /**成功建檔*/
-                singleBatchQueryParams.setStatus("DONE");
-                filePath = directories.getLocalTempDir()+singleBatchQueryParams.getFile();
-            }
-            postSingleBatchQueryRequest("/batchForm/systemUpdate",singleBatchQueryParams,filePath);
-        }catch (Exception e){
-            log.info("e:{}",e.toString());
-            singleBatchQueryParams.setStatus("SYS_FAILED");
-            postSingleBatchQueryRequest("/batchForm/systemUpdate",singleBatchQueryParams,"");
-        }
-    }
-
-    private boolean getCSVFile(List<IbdTbIhChangeDoorplateHis> IbdTbIhChangeDoorplateHisList,String fileName) {
-        StringBuilder newContent = new StringBuilder();
-        try{
-            for (IbdTbIhChangeDoorplateHis line : IbdTbIhChangeDoorplateHisList) {
-                newContent.append(line.getAddressId()).append(",");
-                newContent.append(line.getHisAdr()).append(",");
-                newContent.append(line.getWgsX()).append(",");
-                newContent.append(line.getWgsY()).append(",");
-                newContent.append(line.getUpdateDt()).append(",");
-                newContent.append(line.getUpdateType()).append(",").append("\n");
-            }
-            log.info("newContent:{}",newContent);
-                String filePath =  directories.getLocalTempDir()+ fileName;
-                saveByteArrayToFile(newContent.toString().getBytes(),filePath);
-                return true;
-        } catch (Exception e) {
-            log.info("e:{}",e.toString());
-            return false;
-        }
-    }
-
-    private static void saveByteArrayToFile(byte[] byteArray, String filePath) throws IOException {
-        File file = new File(filePath);
-        try (FileOutputStream fos = new FileOutputStream(file)) {
-            FileCopyUtils.copy(byteArray, fos);
-        }
-    }
-
-
-    public int postSingleBatchQueryRequest(String action, Object params, String filePath) throws IOException {
-        String targerUrl = apServerComponent.getTargetUrl() + action;
-        log.info("targetUrl:{}",targerUrl);
-
-        File file = null;
-        String fileName ="";
-        if(!"".equals(filePath)){
-            file = new File(filePath);
-            if (file.exists()) {
-                fileName = String.valueOf(Path.of(filePath).getFileName());
-                log.info("fileName:{}",fileName);
-            }
-        }
-
-        URL url = new URL(targerUrl);
-        HttpURLConnection con = (HttpURLConnection) url.openConnection();
-        con.setRequestMethod("PUT");
-
-        /**necessay**/
-        con.setDoOutput(true);
-        /**necessay**/
-        String boundary = UUID.randomUUID().toString();
-        con.setRequestProperty("Content-Type", "multipart/form-data; boundary=" + boundary);
-        Path path = Path.of(apServerComponent.getToken());
-        String token = Files.readString(path, StandardCharsets.UTF_8);
-        con.setRequestProperty("Authorization",token);
-//      con.setRequestProperty("Authorization","Bearer eyJhbGciOiJSUzI1NiJ9.eyJ1c2VyIjoie30iLCJqdGkiOiJNVE0yTmpRMk9HWXRPRGcyWWkwME9UTXhMV0UyWlRRdE9URmlNelE1WlRjek5ETTAiLCJleHAiOjE3Mzc2MDE4MzJ9.3ghp8wCHziA6Az9UpS8ssL1d_JB5apN-3pbIV28BWx3bOK-FjRGA9676-EDpqhXrth_Sqln_TFd4wT0RGJ4V1M0RtKXj3EMpFBBV0otdAsgZLm0JcK7LjUrXmWvyfsBcasnHQ83rMo4hE4GeBgXlrhPUlRxnPcVbk4UrVkaMtxyngDfkGpInPJokUWzrScgo7TDA-aKmodw2eZbxYPjGTw1fzXTYHpJC4VNyAYbeGOTd9uMh-cCAyyYMsw__JmkQOAYPpKLnHdyHSb6C8ezxAZJNrI5Rpg4cG0ousXh694IXmixI_R7Q1nVBMFl7GG946fgTO9twiqhuaB64beUILg");
-
-        try (OutputStream out = con.getOutputStream();
-             PrintWriter writer = new PrintWriter(new OutputStreamWriter(out, "UTF-8"), true)) {
-//            writer.append("--").append(boundary).append("\r\n");
-//            writer.append("Content-Disposition: form-data; name=\"id\"\r\n\r\n");
-//            writer.append(params.getBATCH_ID()).append("\r\n");
-
-//            writer.append("--").append(boundary).append("\r\n");
-//            writer.append("Content-Disposition: form-data; name=\"originalFileId\"\r\n\r\n");
-//            writer.append(params.getBATCHFORM_ORIGINAL_FILE_ID()).append("\r\n");
-
-//            writer.append("--").append(boundary).append("\r\n");
-//            writer.append("Content-Disposition: form-data; name=\"processedCounts\"\r\n\r\n");
-//            writer.append("0").append("\r\n");
-
-//            writer.append("--").append(boundary).append("\r\n");
-//            writer.append("Content-Disposition: form-data; name=\"status\"\r\n\r\n");
-//            writer.append(params.getStatus()).append("\r\n");
-            String content = getContent(boundary, params);
-            writer.append(content);
-
-            if (!"".equals(fileName)) {
-                // 如果檔案名稱非空，則將檔案內容寫入請求主體
-                writer.append("--").append(boundary).append("\r\n");
-                writer.append("Content-Disposition: form-data; name=\"file\"; filename=\"").append(fileName).append("\"\r\n");
-                writer.append("Content-Type: application/zip\r\n\r\n");
-                writer.flush();
-                log.info("filePath:{}",filePath);
-                Files.copy(file.toPath(), out);
-                out.flush();
-                writer.append("\r\n");
-            }
-            writer.append("--").append(boundary).append("--").append("\r\n");
-            writer.flush();
-        }
-
-
-        int responseCode = con.getResponseCode();
-        String responseMessage = con.getResponseMessage();
-        log.info("Response Code: " + responseCode);
-        log.info("Response Message: " + responseMessage);
-
-        con.disconnect();
-        return responseCode;
-    }
-
-    public String getContent(String boundary,Object params) {
-        StringBuilder content = new StringBuilder();
-        if (params != null) {
-            Method[] methods = params.getClass().getMethods();
-            for (Method method : methods) {
-                if (method.getName().startsWith("get") && !method.getName().equals("getClass")) {
-                    try {
-                        /**調用getter**/
-                        Object value = method.invoke(params);
-                        if (value != null && !"".equals(value)) {
-                            if (method.getName().indexOf("File") == 3) {
-                                continue;
-                            }else{
-                                content.append("--").append(boundary).append("\r\n");
-                                String name = method.getName().substring(3);
-                                name = Character.toLowerCase(name.charAt(0)) + name.substring(1);
-                                String format = String.format("Content-Disposition: form-data; name=\"%s\"\r\n\r\n",name);
-                                content.append(format);
-                                content.append(value).append("\r\n");
-                            }
-                        }
-                    } catch (Exception e) {
-                        throw new MoiException("url解析錯誤 " + method.getName() + ": " + e.getMessage(), e);
-                    }
-                }
-            }
-        }
-        log.info("content:{}",content);
-        return content.toString();
-    }
-
-
-
-    public List<String> CSVReader(MultipartFile file) throws IOException {
-        try{
-            HashSet<String> addressIdSet = new HashSet<String>();
-            CSVReader csvReader = new CSVReader(new InputStreamReader(file.getInputStream()));
-            String[] line;
-            /**跳過表投**/
-            csvReader.readNext();
-            while ((line = csvReader.readNext()) != null) {
-                /**跳過空行**/
-                if (!isNullOrEmpty(line)) {
-                    log.info("line:{}",line[0]);
-                    /**欄位數符合**/
-                    addressIdSet.add(line[0]);
-                }
-            }
-            return addressIdSet.stream().toList();
-        }catch (Exception e){
-            log.info("e:{}",e.toString());
-            return null;
-        }
-    }
-
-
-    /**判断陣列是否為空或元素都為空字串**/
-    private boolean isNullOrEmpty(String[] array) {
-        return array == null || Arrays.stream(array).allMatch(String::isEmpty);
     }
 }
